+++ starsky/starskytest/starsky.foundation.platform/Middleware/ContentSecurityPolicyMiddlewareTest.cs
@@ -61,5 +61,32 @@ namespace starskytest.starsky.foundation.platform.Middleware
 			Assert.AreEqual(true,csp.Contains("default-src"));
 			Assert.AreEqual(true,csp.Contains("wss://"));
 		}
+		
+		[TestMethod]
+		public async Task ContentSecurityPolicyMiddlewareTest_invoke_otherTypes()
+		{
+			// Arrange
+			var httpContext = new DefaultHttpContext();
+			httpContext.Request.Scheme = "http";
+			var authMiddleware = new ContentSecurityPolicyMiddleware((innerHttpContext) => Task.FromResult(0));
+
+			// Act
+			await authMiddleware.Invoke(httpContext);
+			
+			// test
+			var referrerPolicy = httpContext.Response.Headers["Referrer-Policy"].ToString();
+			Assert.AreEqual( "no-referrer",referrerPolicy);
+			
+			var frameOptions = httpContext.Response.Headers["X-Frame-Options"].ToString();
+			Assert.AreEqual( "DENY",frameOptions);
+			
+			// X-Xss-Protection
+			var xssProtection = httpContext.Response.Headers["X-Xss-Protection"].ToString();
+			Assert.AreEqual( "1; mode=block",xssProtection);
+
+			// X-Content-Type-Options
+			var contentTypeOptions = httpContext.Response.Headers["X-Content-Type-Options"].ToString();
+			Assert.AreEqual( "nosniff",contentTypeOptions);
+		}
 	}
 }